+++ server/writer/gzip.go
@@ -2,6 +2,7 @@ package writer
 
 import (
 	"bufio"
+	"bytes"
 	"compress/gzip"
 	"fmt"
 	"net"
@@ -15,6 +16,8 @@ const (
 	ContentLengthHeader   = "Content-Length"
 	GzipName              = "gzip"
 	VaryHeader            = "Vary"
+
+	minCompressBodyLength = 60
 )
 
 var (
@@ -24,22 +27,58 @@ var (
 )
 
 type Gzip struct {
-	enabled bool
-	rw      http.ResponseWriter
-	w       *gzip.Writer
-	written int64
+	buffer     *bytes.Buffer
+	enabled    bool
+	headerSent bool
+	statusCode int
+	writeErr   error
+	rw         http.ResponseWriter
+	w          *gzip.Writer
 }
 
 func NewGzipWriter(rw http.ResponseWriter, header http.Header) *Gzip {
-	gw := gzip.NewWriter(rw)
 	return &Gzip{
+		buffer:  bytes.NewBuffer(nil),
 		enabled: clientSupportsGZ.MatchString(header.Get(AcceptEncodingHeader)),
 		rw:      rw,
-		w:       gw,
+		w:       gzip.NewWriter(rw),
 	}
 }
 
+// Write fills a small buffer first to determine if a compression is required or not.
 func (g *Gzip) Write(p []byte) (n int, err error) {
+	b := p[:]
+	bytesLen := len(p)
+	bufLen := g.buffer.Len()
+
+	if bufLen < minCompressBodyLength {
+		limit := minCompressBodyLength - bufLen
+
+		if bytesLen < limit {
+			return g.buffer.Write(b)
+		}
+
+		// Fill the buffer at least to minCompressBodyLength size.
+		if _, err = g.buffer.Write(b); err != nil {
+			return 0, err
+		}
+
+		b = g.buffer.Bytes()
+	}
+
+	g.writeHeader()
+
+	n, err = g.write(b)
+	if err != nil {
+		return n, err
+	} else if bufLen < minCompressBodyLength && bytesLen != (n-bufLen) {
+		return 0, fmt.Errorf("invalid write result")
+	}
+
+	return bytesLen, err
+}
+
+func (g *Gzip) write(p []byte) (n int, err error) {
 	if g.enabled {
 		return g.w.Write(p)
 	}
@@ -47,9 +86,26 @@ func (g *Gzip) Write(p []byte) (n int, err error) {
 }
 
 func (g *Gzip) Close() (err error) {
+	if g.writeErr != nil {
+		return g.writeErr
+	}
+
+	if g.buffer.Len() < minCompressBodyLength {
+		g.enabled = false
+		g.writeHeader()
+
+		_, err = g.write(g.buffer.Bytes())
+		if err != nil {
+			return err
+		}
+	}
+
+	g.writeHeader()
+
 	if g.enabled && g.w != nil {
 		err = g.w.Close()
 	}
+
 	return err
 }
 
@@ -58,17 +114,37 @@ func (g *Gzip) Header() http.Header {
 }
 
 func (g *Gzip) WriteHeader(statusCode int) {
-	g.rw.Header().Add(VaryHeader, AcceptEncodingHeader)
+	g.statusCode = statusCode
+}
+
+func (g *Gzip) writeHeader() {
+	if g.headerSent {
+		return
+	}
+
+	g.headerSent = true
+
+	if g.buffer.Len() >= minCompressBodyLength {
+		g.rw.Header().Add(VaryHeader, AcceptEncodingHeader)
+	}
 
 	if g.enabled {
 		g.rw.Header().Del(ContentLengthHeader)
 		g.rw.Header().Set(ContentEncodingHeader, GzipName)
 	}
 
-	g.rw.WriteHeader(statusCode)
+	g.rw.WriteHeader(g.statusCode)
 }
 
 func (g *Gzip) Flush() {
+	if l := g.buffer.Len(); l < minCompressBodyLength {
+		// We have to wait for minCompressBodyLength bytes to be
+		// able to determine, if we enable GZIP compression or not.
+		return
+	}
+
+	g.writeHeader()
+
 	if g.enabled && g.w != nil {
 		_ = g.w.Flush()
 	}