+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Unreleased changes are available as `avenga/couper:edge` container.
 * **Added**
   * Run option `-accept-forwarded-url` and setting `accept_forwarded_url` to accept `proto`, `host`, or `port` from `X-Forwarded-Proto`, `X-Forwarded-Host` or `X-Forwarded-Port` request headers, affecting new client request variables `request.url`, `request.origin`, `request.protocol`, `request.host` and `request.port` ([#255](https://github.com/avenga/couper/pull/255))
 
+* **Fixed**
+  * No GZIP compression for small response bodies ([#186](https://github.com/avenga/couper/issues/186))
+
 * [**Beta**](./docs/BETA.md)
   * OAuth2 Authorization Code Grant Flow: [`beta_oauth2 {}` block](./docs/REFERENCE.md#oauth2-ac-block-beta);  [`beta_oauth_authorization_url()`](./docs/REFERENCE.md#functions), [`beta_oauth_code_verifier()`](./docs/REFERENCE.md#functions) and [`beta_oauth_csrf_token()` functions](./docs/REFERENCE.md#functions) ([#247](https://github.com/avenga/couper/pull/247))